evaluate_instruction: Deal with more 16 bit unofficial opcodes

+++ emu-core/src/evaluate_instruction.c
@@ -804,8 +804,8 @@ STATIC_INLINE void lax_absolute(State *state, uint8_t val) {
 }
 
 STATIC_INLINE void lax_absolute_y(State *state, uint16_t val) {
-	state->cpu.a = val;
-	state->cpu.x = val;
+	state->cpu.a = (uint8_t)val;
+	state->cpu.x = (uint8_t)val;
 	// Update flags
 	state->cpu.p.Z = (val == 0);
 	state->cpu.p.N = ((val & 0x80) != 0);
@@ -842,9 +842,9 @@ STATIC_INLINE void sax_zero_page_y(State *state, uint8_t val) {
 	state->cpu.pc += 2;
 }
 
-STATIC_INLINE void sax_absolute(State *state, uint8_t val) {
+STATIC_INLINE void sax_absolute(State *state, uint16_t val) {
 	uint8_t result = state->cpu.a & state->cpu.x;
-	state_set_mem(state, val, result);
+	state_set_mem(state, (uint8_t)val, result);
 	state->cpu.pc += 2;
 }
 
@@ -878,9 +878,9 @@ STATIC_INLINE void dcp_zero_page_x(State *state, uint8_t val) {
 	state->cpu.pc += 2;
 }
 
-STATIC_INLINE void dcp_absolute(State *state, uint8_t val) {
-	uint8_t result = val - 1;
-	state_set_mem(state, val, result);
+STATIC_INLINE void dcp_absolute(State *state, uint16_t val) {
+	uint8_t result = (uint8_t)val - 1;
+	state_set_mem(state, (uint8_t)val, result);
 
 	// Compare
 	uint8_t temp   = state->cpu.a - result;
@@ -964,9 +964,9 @@ STATIC_INLINE void isc_zero_page_x(State *state, uint8_t val) {
 	state->cpu.pc += 2;
 }
 
-STATIC_INLINE void isc_absolute(State *state, uint8_t val) {
-	uint8_t result = val + 1;
-	state_set_mem(state, val, result);
+STATIC_INLINE void isc_absolute(State *state, uint16_t val) {
+	uint8_t result = (uint8_t)val + 1;
+	state_set_mem(state, (uint8_t)val, result);
 
 	// Subtract with borrow
 	uint8_t temp   = state->cpu.a - result - (1 - state->cpu.p.C);
@@ -978,7 +978,7 @@ STATIC_INLINE void isc_absolute(State *state, uint8_t val) {
 }
 
 STATIC_INLINE void isc_absolute_x(State *state, uint16_t val) {
-	uint8_t result = val + 1;
+	uint8_t result = (uint8_t)val + 1;
 	state_set_mem(state, val, result);
 
 	// Subtract with borrow
@@ -991,7 +991,7 @@ STATIC_INLINE void isc_absolute_x(State *state, uint16_t val) {
 }
 
 STATIC_INLINE void isc_absolute_y(State *state, uint16_t val) {
-	uint8_t result = val + 1;
+	uint8_t result = (uint8_t)val + 1;
 	state_set_mem(state, val, result);
 
 	// Subtract with borrow
@@ -1083,9 +1083,9 @@ STATIC_INLINE void rla_absolute(State *state, uint8_t val) {
 	state->cpu.pc += 2;
 }
 
-STATIC_INLINE void rla_absolute_x(State *state, uint8_t val) {
+STATIC_INLINE void rla_absolute_x(State *state, uint16_t val) {
 	// Rotate left
-	uint8_t carry  = (val & 0x80) ? 1 : 0;
+	uint8_t carry  = ((uint8_t)val & 0x80) ? 1 : 0;
 	uint8_t result = (uint8_t) ((val << 1) | (uint8_t) state->cpu.p.C);
 	state->cpu.p.C = carry;
 
@@ -1096,7 +1096,7 @@ STATIC_INLINE void rla_absolute_x(State *state, uint8_t val) {
 	state->cpu.p.Z = (result == 0);
 	state->cpu.p.N = ((result & 0x80) != 0);
 
-	state_set_mem(state, val, result);
+	state_set_mem(state, (uint8_t)val, result);
 	state->cpu.a = result;
 	state->cpu.pc += 2;
 }
@@ -1189,9 +1189,9 @@ STATIC_INLINE void rra_zero_page_x(State *state, uint8_t val) {
 	state->cpu.pc += 2;
 }
 
-STATIC_INLINE void rra_absolute(State *state, uint8_t val) {
+STATIC_INLINE void rra_absolute(State *state, uint16_t val) {
 	// Rotate right
-	uint8_t carry  = (val & 0x01) ? 0x80 : 0;
+	uint8_t carry  = ((uint8_t)val & 0x01) ? 0x80 : 0;
 	uint8_t result = (uint8_t) ((val >> 1) | ((uint8_t) state->cpu.p.C << 7));
 	state->cpu.p.C = carry;
 
@@ -1206,9 +1206,9 @@ STATIC_INLINE void rra_absolute(State *state, uint8_t val) {
 	state->cpu.pc += 2;
 }
 
-STATIC_INLINE void rra_absolute_x(State *state, uint8_t val) {
+STATIC_INLINE void rra_absolute_x(State *state, uint16_t val) {
 	// Rotate right
-	uint8_t carry  = (val & 0x01) ? 0x80 : 0;
+	uint8_t carry  = ((uint8_t)val & 0x01) ? 0x80 : 0;
 	uint8_t result = (uint8_t) ((val >> 1) | ((uint8_t) state->cpu.p.C << 7));
 	state->cpu.p.C = carry;
 
@@ -1218,7 +1218,7 @@ STATIC_INLINE void rra_absolute_x(State *state, uint8_t val) {
 	state->cpu.p.Z = (temp == 0);
 	state->cpu.p.N = ((temp & 0x80) != 0);
 
-	state_set_mem(state, val, result);
+	state_set_mem(state, (uint8_t)val, result);
 	state->cpu.a = temp;
 	state->cpu.pc += 2;
 }
@@ -1641,8 +1641,8 @@ STATIC_INLINE void shx_absolute_y(State *state, uint16_t) {
 
 STATIC_INLINE void ahx_absolute_y(State *state, uint16_t val) {
 	// Store A and X registers with high byte of adress
-	uint8_t adr_low  = val;
-	uint8_t adr_high = (val + 1) & 0xFF;
+	uint8_t adr_low  = (uint8_t)val;
+	uint8_t adr_high = ((uint8_t)val + 1) & 0xFF;
 
 	// Store (A & X) in memory
 	state_set_mem(state, (uint16_t) (adr_low | (adr_high << 8)), state->cpu.a & state->cpu.x);
